auth_service/pyproject.toml
[project]
name = "auth_service"
version = "0.1.0"
description = "..."
authors = [{name = "Jahongir", email = "jahongir.hodjaniyazov@gmail.com"}]
requires-python = ">=3.13"
dependencies = [
    "fastapi ==0.116.1",
    "uvicorn ==0.35.0",
    "sqlalchemy ==2.0.41",
    "pydantic-settings ==2.10.1",
    "aiokafka ==0.12.0",
    "asyncpg ==0.30.0",
    "greenlet ==3.2.3",
    "passlib ==1.7.4",
    "bcrypt ==4.0.1",
    "python-jose ==3.5.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build_backend = "setuptools.build_meta"